refactor: split bridge.rs into bridge/mod.rs and bridge/types.rs submodule (#800)
